+++ src/Enums/AreaType.php
@@ -10,6 +10,7 @@ use MyCLabs\Enum\Enum;
  * @method static AreaType UNKNOWN()
  * @method static AreaType MUNICIPALITY()
  * @method static AreaType SUBMUNICIPALITY()
+ * @method static AreaType ROAD_AUTHORITY()
  * @method static AreaType TUNNEL_AUTHORITY()
  * @method static AreaType PROVINCE()
  */
@@ -18,6 +19,7 @@ final class AreaType extends Enum
     private const UNKNOWN = 'UNKNOWN';
     private const MUNICIPALITY = 'MUNICIPALITY';
     private const SUBMUNICIPALITY = 'SUBMUNICIPALITY';
+    private const ROAD_AUTHORITY = 'ROAD_AUTHORITY';
     private const TUNNEL_AUTHORITY = 'TUNNEL_AUTHORITY';
     private const PROVINCE = 'PROVINCE';
 
@@ -27,6 +29,7 @@ final class AreaType extends Enum
             'UNKNOWN' => '',
             'MUNICIPALITY' => 'Gemeente',
             'SUBMUNICIPALITY' => 'Gemeente',
+            'ROAD_AUTHORITY' => 'Wegendistrict',
             'TUNNEL_AUTHORITY' => 'Tunnelbeheerder',
             'PROVINCE' => 'Provincie',
         ][$this->getKey()];